each option should have its own tooltip
