+++ app/components/forms/MultiSelectControl/Option.js
@@ -1,13 +1,64 @@
-import React from 'react';
+import React, { useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { palette } from 'styled-theme';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, injectIntl, intlShape } from 'react-intl';
+import {
+  Box, Button, Drop, Text,
+} from 'grommet';
+import { CircleQuestion } from 'grommet-icons';
 
-import { lowerCase } from 'utils/string';
+import { lowerCase, truncateText } from 'utils/string';
 import appMessage from 'utils/app-message';
+import IndeterminateCheckbox from 'components/forms/IndeterminateCheckbox';
 
 import messages from './messages';
+
+const Styled = styled((p) => (
+  <Box fill="horizontal" direction="row" align="center" {...p} />
+))`
+  width: 100%;
+  line-height: 1.3;
+  font-size: 0.8em;
+  @media (min-width: ${(props) => props.theme.breakpoints.small}) {
+    border-bottom: 1px solid ${palette('light', 1)};
+    font-size: 1em;
+  }
+  @media print {
+    font-size: ${(props) => props.theme.sizes.print.default};
+  }
+`;
+
+const CheckboxWrapper = styled((p) => (
+  <Box
+    fill="vertical"
+    flex={{ shrink: 0 }}
+    align="center"
+    {...p}
+  />
+))`
+  width: 40px;
+`;
+
+const OptionLabel = styled((p) => <Box pad={{ vertical: 'small' }} fill as="label" {...p} />)`
+  vertical-align: middle;
+  cursor: pointer;
+  border-right: ${(props) => (props.changedToChecked || props.changedToUnchecked)
+    ? '0.5em solid'
+    : 'none'
+};
+  border-right-color: ${palette('buttonDefault', 1)};
+`;
+
+const DropContent = styled((p) => (
+  <Box
+    pad="small"
+    background="light-1"
+    {...p}
+  />
+))`
+  max-width: 280px;
+`;
 
 // font-weight: ${(props) => props.bold ? 500 : 'normal'};
 const Label = styled.div`
@@ -38,56 +89,130 @@ const IdSpacer = styled.span`
   color: ${palette('text', 1)};
 `;
 // <Label bold={props.bold} italic={props.isNew}>
-class Option extends React.Component { // eslint-disable-line react/prefer-stateless-function
-  render() {
-    const { intl } = this.context;
-    const {
-      reference, message, label, messagePrefix, isNew, emphasis = false,
-    } = this.props;
+function Option({
+  option,
+  onCheckboxChange,
+  optionId,
+  intl,
+  secondary,
+}) {
+  const [showInfo, setShowInfo] = useState(false);
+  const infoRef = useRef(null);
 
-    let optionLabel;
-    if (message) {
-      optionLabel = messagePrefix
-        ? `${messagePrefix} ${lowerCase(appMessage(intl, message))}`
-        : appMessage(intl, message);
-    } else {
-      optionLabel = label;
-    }
+  const emphasis = option.get('labelEmphasis');
+  const reference = typeof option.get('reference') !== 'undefined' && option.get('reference') !== null ? option.get('reference').toString() : '';
+  const label = option.get('label');
+  const optionInfo = option.get('info');
+  const messagePrefix = option.get('messagePrefix');
+  const message = option.get('message');
+  const isNew = option.get('isNew');
+  // const draft = option.get('draft');
+  const checked = option.get('checked');
+  const isIndeterminate = option.get('isIndeterminate');
 
-    return (
-      <Label emphasis={emphasis}>
-        {reference
-          && <Id>{reference}</Id>
-        }
-        {reference
-          && <IdSpacer />
-        }
-        { optionLabel }
-        {isNew
+  let optionLabel;
+  if (message) {
+    optionLabel = messagePrefix
+      ? `${messagePrefix} ${lowerCase(appMessage(intl, message))}`
+      : appMessage(intl, message);
+  } else {
+    optionLabel = label;
+  }
+
+  return (
+    <Styled
+      changedToChecked={option.get('changedToChecked')}
+      changedToUnchecked={option.get('changedToUnchecked')}
+      secondary={secondary}
+    >
+      <CheckboxWrapper>
+        { isIndeterminate
           && (
-            <New>
-              <FormattedMessage {...messages.new} />
-            </New>
+            <IndeterminateCheckbox
+              id={optionId}
+              checked={checked}
+              onChange={(checkedState) => {
+                onCheckboxChange(checkedState, option);
+              }}
+            />
           )
         }
-      </Label>
-    );
-  }
+        { !isIndeterminate
+          && (
+            <input
+              id={optionId}
+              type="checkbox"
+              checked={checked}
+              onChange={(evt) => {
+                evt.stopPropagation();
+                onCheckboxChange(evt.target.checked, option);
+              }}
+            />
+          )
+        }
+      </CheckboxWrapper>
+      <OptionLabel
+        htmlFor={optionId}
+        changedToChecked={option.get('changedToChecked')}
+        changedToUnchecked={option.get('changedToUnchecked')}
+        secondary={secondary}
+      >
+        <Label emphasis={emphasis}>
+          {reference
+            && <Id>{reference}</Id>
+          }
+          {reference
+            && <IdSpacer />
+          }
+          { optionLabel }
+          {isNew
+            && (
+              <New>
+                <FormattedMessage {...messages.new} />
+              </New>
+            )
+          }
+        </Label>
+      </OptionLabel>
+      {optionInfo && (
+        <Box
+          fill={false}
+          pad={{ horizontal: 'small' }}
+          ref={infoRef}
+          flex={{ shrink: 0 }}
+        >
+          <Button
+            plain
+            icon={<CircleQuestion color={option.get('active') ? 'white' : 'dark-2'} />}
+            fill={false}
+            onMouseOver={() => setShowInfo(true)}
+            onMouseLeave={() => setShowInfo(false)}
+            onFocus={() => setShowInfo(true)}
+            onBlur={() => null}
+            onClick={() => setShowInfo(!showInfo)}
+          />
+        </Box>
+      )}
+      {optionInfo && showInfo && infoRef && (
+        <Drop
+          align={{ top: 'top', right: 'left' }}
+          target={infoRef.current}
+        >
+          <DropContent>
+            <Text size="small">{truncateText(optionInfo, 600, true)}</Text>
+          </DropContent>
+        </Drop>
+      )}
+    </Styled>
+  );
 }
 
 Option.propTypes = {
-  label: PropTypes.string,
-  message: PropTypes.string,
-  messagePrefix: PropTypes.string,
-  reference: PropTypes.string,
-  info: PropTypes.string,
-  // draft: PropTypes.bool,
-  isNew: PropTypes.bool,
-  emphasis: PropTypes.bool,
+  optionId: PropTypes.string,
+  option: PropTypes.object,
+  secondary: PropTypes.bool,
+  onCheckboxChange: PropTypes.func,
+  intl: intlShape.isRequired,
 };
 
-Option.contextTypes = {
-  intl: PropTypes.object.isRequired,
-};
-
-export default Option;
+export default injectIntl(Option);
